+++ state/executor.go
@@ -203,7 +203,7 @@ func (e *BlockExecutor) updateState(state types.State, block *types.Block, abciR
 				}
 			}
 			// Change results from this height but only applies to the next next height.
-			lastHeightValSetChanged = block.Height() + 1 + 1
+			lastHeightValSetChanged = int64(block.Height()) + 1 + 1
 		}
 
 		if len(nValSet.Validators) > 0 {
@@ -218,7 +218,7 @@ func (e *BlockExecutor) updateState(state types.State, block *types.Block, abciR
 		Version:         state.Version,
 		ChainID:         state.ChainID,
 		InitialHeight:   state.InitialHeight,
-		LastBlockHeight: block.Height(),
+		LastBlockHeight: int64(block.Height()),
 		LastBlockTime:   block.Time(),
 		LastBlockID: cmtypes.BlockID{
 			Hash: cmbytes.HexBytes(block.Hash()),
@@ -270,10 +270,10 @@ func (e *BlockExecutor) validate(state types.State, block *types.Block) error {
 		block.SignedHeader.Version.Block != state.Version.Consensus.Block {
 		return errors.New("block version mismatch")
 	}
-	if state.LastBlockHeight <= 0 && block.Height() != state.InitialHeight {
+	if state.LastBlockHeight <= 0 && int64(block.Height()) != state.InitialHeight {
 		return errors.New("initial block height mismatch")
 	}
-	if state.LastBlockHeight > 0 && block.Height() != state.LastBlockHeight+1 {
+	if state.LastBlockHeight > 0 && int64(block.Height()) != state.LastBlockHeight+1 {
 		return errors.New("block height mismatch")
 	}
 	if !bytes.Equal(block.SignedHeader.AppHash[:], state.AppHash[:]) {
@@ -341,7 +341,7 @@ func (e *BlockExecutor) execute(ctx context.Context, state types.State, block *t
 		}
 
 	}
-	endBlockRequest := abci.RequestEndBlock{Height: block.Height()}
+	endBlockRequest := abci.RequestEndBlock{Height: int64(block.Height())}
 	abciResponses.EndBlock, err = e.proxyApp.EndBlockSync(endBlockRequest)
 	if err != nil {
 		return nil, err
@@ -375,13 +375,13 @@ func (e *BlockExecutor) publishEvents(resp *cmstate.ABCIResponses, block *types.
 	for _, ev := range abciBlock.Evidence.Evidence {
 		err = multierr.Append(err, e.eventBus.PublishEventNewEvidence(cmtypes.EventDataNewEvidence{
 			Evidence: ev,
-			Height:   block.Height(),
+			Height:   int64(block.Height()),
 		}))
 	}
 	for i, dtx := range resp.DeliverTxs {
 		err = multierr.Append(err, e.eventBus.PublishEventTx(cmtypes.EventDataTx{
 			TxResult: abci.TxResult{
-				Height: block.Height(),
+				Height: int64(block.Height()),
 				Index:  uint32(i),
 				Tx:     abciBlock.Data.Txs[i],
 				Result: *dtx,
